feat: 레스토랑 클래스 생성
