+++ src/types/category.ts
@@ -1,0 +1,3 @@
+type Category = '한식' | '중식' | '일식' | '아시안' | '양식' | '기타';
+
+export default Category;
