+++ src/types/index.ts
@@ -1,0 +1,2 @@
+export { default as RestaurantType } from './restaurant';
+export { default as CategoryType } from './category';
